+++ src/test/lit-element_test.ts
@@ -1310,9 +1310,7 @@ suite('LitElement', () => {
         this.changedProperties = changedProperties;
       }
 
-      render() {
-        return html`${this.id}-${this.name}-${this.title}-${this.foo}`;
-      }
+      render() { return html`${this.id}-${this.title}-${this.foo}`; }
     }
     customElements.define(generateElementName(), E);
     const el = new E() as any;
@@ -1320,16 +1318,230 @@ suite('LitElement', () => {
     await el.updateComplete;
     el.foo = 'foo';
     el.id = 'id';
-    el.name = 'name';
     el.title = 'title';
     await el.updateComplete;
-    assert.equal(el.shadowRoot!.textContent, 'id-name-title-foo');
+    assert.equal(el.shadowRoot!.textContent, 'id-title-foo');
     assert.equal((window as any).id, el);
+    assert.equal(el.getAttribute('id'), 'id');
     el.id = 'id2';
-    el.name = 'name2';
     await el.updateComplete;
-    assert.equal(el.shadowRoot!.textContent, 'id2-name2-title-foo');
+    assert.equal(el.shadowRoot!.textContent, 'id2-title-foo');
     assert.equal((window as any).id2, el);
+    assert.equal(el.getAttribute('id'), 'id2');
+  });
+
+  test('user accessors correctly wrapped', async () => {
+    class Sup extends LitElement {
+      _supSet?: number;
+      _oldFoo?: any;
+      _foo?: number;
+      static get properties() { return {foo : {type : Number}}; }
+      constructor() {
+        super();
+        this.foo = 0;
+      }
+      set foo(v: number) {
+        this._supSet = (this._supSet || 0) + 1;
+        this._foo = Math.min(v, 10);
+      }
+      get foo(): number { return this._foo as number; }
+      update(changedProperties: PropertyValues) {
+        this._oldFoo = changedProperties.get('foo');
+        super.update(changedProperties);
+      }
+      render() { return html`${this.foo}`; }
+    }
+    customElements.define(generateElementName(), Sup);
+    class Sub extends Sup {
+      _subSet?: number;
+      static get properties() { return {foo : {type : Number}}; }
+      set foo(v: number) {
+        this._subSet = (this._subSet || 0) + 1;
+        super.foo = v;
+      }
+      get foo(): number {
+        const v = super.foo;
+        return v ? Math.floor(v) : v;
+      }
+    }
+    customElements.define(generateElementName(), Sub);
+
+    const sup = new Sup();
+    container.appendChild(sup);
+    await sup.updateComplete;
+    assert.equal(sup.foo, 0);
+    assert.equal(sup._oldFoo, undefined);
+    assert.equal(sup._supSet, 1);
+    assert.equal(sup.shadowRoot!.textContent, '0');
+
+    sup.foo = 5;
+    await sup.updateComplete;
+    assert.equal(sup.foo, 5);
+    assert.equal(sup._oldFoo, 0);
+    assert.equal(sup._supSet, 2);
+    assert.equal(sup.shadowRoot!.textContent, '5');
+
+    sup.foo = 20;
+    await sup.updateComplete;
+    assert.equal(sup.foo, 10); // (user getter implements a max of 10)
+    assert.equal(sup._oldFoo, 5);
+    assert.equal(sup._supSet, 3);
+    assert.equal(sup.shadowRoot!.textContent, '10');
+
+    sup.foo = 5;
+    await sup.updateComplete;
+    assert.equal(sup.foo, 5);
+    assert.equal(sup._oldFoo, 10);
+    assert.equal(sup._supSet, 4);
+    assert.equal(sup.shadowRoot!.textContent, '5');
+
+    const sub = new Sub();
+    container.appendChild(sub);
+    await sub.updateComplete;
+    assert.equal(sub.foo, 0);
+    assert.equal(sub._oldFoo, undefined);
+    assert.equal(sub._subSet, 1);
+    assert.equal(sub.shadowRoot!.textContent, '0');
+
+    sub.foo = 5;
+    await sub.updateComplete;
+    assert.equal(sub.foo, 5);
+    assert.equal(sub._oldFoo, 0);
+    assert.equal(sub._subSet, 2);
+    assert.equal(sub.shadowRoot!.textContent, '5');
+
+    sub.foo = 7.5;
+    await sub.updateComplete;
+    assert.equal(sub.foo, 7); // (sub setter rounds down)
+    assert.equal(sub._oldFoo, 5);
+    assert.equal(sub._subSet, 3);
+    assert.equal(sub.shadowRoot!.textContent, '7');
+
+    sub.foo = 20;
+    await sub.updateComplete;
+    assert.equal(sub.foo, 10); // (super user getter maxes at 10)
+    assert.equal(sub._oldFoo, 7);
+    assert.equal(sub._subSet, 4);
+    assert.equal(sub.shadowRoot!.textContent, '10');
+  });
+
+  test('user accessors only using noAccessor', async () => {
+    class E extends LitElement {
+      _updateCount = 0;
+      _foo?: String;
+      _bar?: String;
+      static get properties() {
+        return {
+          foo : {type : String, reflect : true, noAccessor : true},
+          bar : {type : String, reflect : true, noAccessor : true}
+        };
+      }
+      constructor() {
+        super();
+        this.foo = 'defaultFoo';
+        this.bar = 'defaultBar';
+      }
+      set foo(value: string) {
+        const old = this._foo;
+        this._foo = value;
+        this.requestUpdate('foo', old);
+      }
+      get foo() { return this._foo as string; }
+      set bar(value: string) {
+        const old = this._bar;
+        this._bar = value;
+        this.requestUpdate('bar', old);
+      }
+      get bar() { return this._bar as string; }
+      update(changedProperties: PropertyValues) {
+        this._updateCount++;
+        super.update(changedProperties);
+      }
+      render() { return html`${this.foo}-${this.bar}`; }
+    }
+    customElements.define(generateElementName(), E);
+
+    const el = new E();
+    el.foo = 'foo1';
+    document.body.appendChild(el);
+    await el.updateComplete;
+    assert.equal(el.foo, 'foo1');
+    assert.equal(el.bar, 'defaultBar');
+    assert.equal(el.getAttribute('foo'), 'foo1');
+    assert.equal(el.getAttribute('bar'), 'defaultBar');
+    assert.equal(el.shadowRoot!.textContent, 'foo1-defaultBar');
+    assert.equal(el._updateCount, 1);
+
+    el.foo = 'foo2';
+    el.bar = 'bar';
+    await el.updateComplete;
+    assert.equal(el.foo, 'foo2');
+    assert.equal(el.bar, 'bar');
+    assert.equal(el.getAttribute('foo'), 'foo2');
+    assert.equal(el.getAttribute('bar'), 'bar');
+    assert.equal(el.shadowRoot!.textContent, 'foo2-bar');
+    assert.equal(el._updateCount, 2);
+
+    el.foo = 'foo3';
+    await el.updateComplete;
+    assert.equal(el.foo, 'foo3');
+    assert.equal(el.getAttribute('foo'), 'foo3');
+    assert.equal(el.getAttribute('bar'), 'bar');
+    assert.equal(el.shadowRoot!.textContent, 'foo3-bar');
+    assert.equal(el._updateCount, 3);
+  });
+
+  test('attributeChangedCallback based rendering', async () => {
+    class E extends LitElement {
+      _updateCount = 0;
+      static get properties() {
+        return {
+          foo : {type : String, noAccessor : true},
+          bar : {type : String, noAccessor : true}
+        };
+      }
+      set foo(value: string|null) { this.setAttribute('foo', value as string); }
+      get foo() { return this.getAttribute('foo') || 'defaultFoo'; }
+      set bar(value: string|null) { this.setAttribute('bar', value as string); }
+      get bar() { return this.getAttribute('bar') || 'defaultBar'; }
+      attributeChangedCallback(name: string, old: null|string,
+                               _value: null|string) {
+        this.requestUpdate(name, old);
+      }
+      update(changedProperties: PropertyValues) {
+        this._updateCount++;
+        super.update(changedProperties);
+      }
+      render() { return html`${this.foo}-${this.bar}`; }
+    }
+    customElements.define(generateElementName(), E);
+
+    const el = new E();
+    el.foo = 'foo1';
+    document.body.appendChild(el);
+    await el.updateComplete;
+    assert.equal(el.foo, 'foo1');
+    assert.equal(el.bar, 'defaultBar');
+    assert.equal(el.shadowRoot!.textContent, 'foo1-defaultBar');
+    assert.equal(el._updateCount, 1);
+
+    el.foo = 'foo2';
+    el.bar = 'bar';
+    await el.updateComplete;
+    assert.equal(el.foo, 'foo2');
+    assert.equal(el.bar, 'bar');
+    assert.equal(el.getAttribute('foo'), 'foo2');
+    assert.equal(el.getAttribute('bar'), 'bar');
+    assert.equal(el.shadowRoot!.textContent, 'foo2-bar');
+    assert.equal(el._updateCount, 2);
+
+    el.foo = 'foo3';
+    await el.updateComplete;
+    assert.equal(el.foo, 'foo3');
+    assert.equal(el.getAttribute('foo'), 'foo3');
+    assert.equal(el.getAttribute('bar'), 'bar');
+    assert.equal(el.shadowRoot!.textContent, 'foo3-bar');
+    assert.equal(el._updateCount, 3);
   });
 
   test(